+++ Tests/XcodeProjTests/Objects/BuildPhase/PBXShellScriptBuildPhaseTests.swift
@@ -24,6 +24,28 @@ final class PBXShellScriptBuildPhaseTests: XCTestCase {
         }
     }
 
+    func test_write_alwaysOutOfDate() throws {
+        // Given
+        let alwaysOutOfDateNotPresent = PBXShellScriptBuildPhase()
+        let alwaysOutOfDateFalse = PBXShellScriptBuildPhase(alwaysOutOfDate: false)
+        let alwaysOutOfDateTrue = PBXShellScriptBuildPhase(alwaysOutOfDate: true)
+        let proj = PBXProj.fixture()
+
+        // When
+        guard
+            case let .dictionary(valuesWhenNotPresent) = try alwaysOutOfDateNotPresent.plistKeyAndValue(proj: proj, reference: "ref").value,
+            case let .dictionary(valuesWhenFalse) = try alwaysOutOfDateFalse.plistKeyAndValue(proj: proj, reference: "ref").value,
+            case let .dictionary(valuesWhenTrue) = try alwaysOutOfDateTrue.plistKeyAndValue(proj: proj, reference: "ref").value else {
+            XCTFail("Plist should contain dictionary")
+            return
+        }
+
+        // Then
+        XCTAssertFalse(valuesWhenNotPresent.keys.contains("alwaysOutOfDate"))
+        XCTAssertFalse(valuesWhenFalse.keys.contains("alwaysOutOfDate"))
+        XCTAssertEqual(valuesWhenTrue["alwaysOutOfDate"], "1")
+    }
+  
     func test_write_dependencyFile() throws {
         let discoveryPath = "$(DERIVED_FILE_DIR)/target.d"
         let discovery = PBXShellScriptBuildPhase(dependencyFile: discoveryPath)